# Optimize JSON.dump argument parsing